[OSD][Tests] add test subject to app title for app analytics
Using a test subject we can find the specific element instead
of trying to search the DOM for the class and hope the class
is the right class that contains the element we are looking for.

We can search the dom for this specific test subject and actually
make the test runner wait until this test subject exists before
executing the tests.

Issue related:
https://github.com/opensearch-project/observability/issues/679

Signed-off-by: Kawika Avilla <kavilla414@gmail.com>